+++ src/main.rs
@@ -4,7 +4,7 @@ mod pull;
 
 fn main() {
     let cli = cli::Cli::parse();
-    if let Err(e) = sup::run_sup(cli.r#continue, cli.abort, cli.version) {
+    if let Err(e) = sup::run_sup(cli.r#continue, cli.abort, cli.version, cli.message) {
         println!("Error: {e}");
         std::process::exit(1);
     }